Refactor de protocolo. Merge entre Match, Game Manager y Game

+++ libs/sub_server/Match.h
@@ -8,6 +8,7 @@
 
 #include <map>
 #include <memory>
+#include <functional>
 #include "box2d/box2d.h"
 #include "Car.h"
 #include "MyContactListener.h"
@@ -20,9 +21,9 @@ private:
     std::string name;
     b2World world;
     Ball* ball;
-    std::map<int, Car*> players;
-    std::map<int, Car*> scorers;
-    std::map<int, Car*> assistants;
+    std::map<uint8_t , Car*> players;
+    std::map<uint8_t , Car*> scorers;
+    std::map<uint8_t , Car*> assistants;
     int playersConnected;
     int playersRequired;
     int goalsLocal;
@@ -34,15 +35,15 @@ private:
 public:
     Match(std::string name, int required);
     ~Match();
-    void addPlayer(int &id);
+    void addPlayer(uint8_t &id);
     float carsInfo();
-    void moveRight(int &id, std::function<void(ActionUpdate * )> function);
+    void moveRight(uint8_t &id, std::function<void(ActionUpdate * )> function);
     void info();
-    void moveLeft(int &id, std::function<void(ActionUpdate * )> function);
-    void jump(int &id, std::function<void(ActionUpdate * )> function);
+    void moveLeft(uint8_t &id, std::function<void(ActionUpdate * )> function);
+    void jump(uint8_t &id, std::function<void(ActionUpdate * )> function);
     void checkGoals();
     void step();
-    void updateGame(int &id);
+    void updateGame(uint8_t &id);
     int visit();
     int local();
